fix: lazyload using component
fix: export

fix: test?
