+++ packages/dm-core-plugins/src/form/fields/BooleanField.test.tsx
@@ -1,6 +1,5 @@
 import { fireEvent, render, screen, waitFor } from '@testing-library/react'
 import userEvent from '@testing-library/user-event'
-import React from 'react'
 import { Form } from '../components/Form'
 import { mockBlueprintGet, wrapper } from '../test-utils'
