+++ packages/dm-core-plugins/src/form/fields/StringField.test.tsx
@@ -6,7 +6,6 @@ import {
   waitFor,
 } from '@testing-library/react'
 import userEvent from '@testing-library/user-event'
-import React from 'react'
 import { Form } from '../components/Form'
 import { mockBlueprintGet, wrapper } from '../test-utils'
 import { TFormProps } from '../types'
