+++ packages/dm-core/src/components/index.tsx
@@ -4,6 +4,7 @@ export * from './DynamicTable'
 export * from './EntityView'
 export * from './JsonView'
 export * from './Loading'
+export * from './LazyLoad'
 export * from './NewEntityButton'
 export * from './DeleteHardButton'
 export * from './AddRowButton'
