+++ packages/dm-core/src/index.tsx
@@ -11,6 +11,7 @@ export * from './components/Dropdown'
 export * from './components/DeleteSoftButton'
 export * from './components/MediaContent'
 export * from './components/TreeView'
+export * from './components/LazyLoad'
 export * from './components/DeleteHardButton'
 export * from './context/ApplicationContext'
 export * from './context/DMSSContext'
